requirements.txt
beautifulsoup4>=4.14.0
python-dateutil>=2.9.0
requests>=2.33.0
icalendar>=6.1.1
